# lib/api-zod/src/generated/types/audibleBook.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AudibleBook {
  asin: string;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  authors: string[];
  narrators: string[];
  /** @nullable */
  coverUrl?: string | null;
  /** @nullable */
  runtimeMinutes?: number | null;
  /** @nullable */
  purchaseDate?: string | null;
  /** @nullable */
  seriesTitle?: string | null;
  /** @nullable */
  seriesPosition?: string | null;
  /** @nullable */
  releaseDate?: string | null;
  /** @nullable */
  description?: string | null;
  /** available | downloaded | downloading */
  status: string;
}