# lib/api-zod/src/generated/types/authStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthStatus {
  authenticated: boolean;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  marketplace?: string | null;
  /** @nullable */
  expiresAt?: string | null;
}