# lib/api-zod/src/generated/types/authUrlResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUrlResponse {
  url: string;
  codeVerifier: string;
  marketplace: string;
}