# lib/api-zod/src/generated/types/downloadJob.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DownloadJob {
  id: string;
  asin: string;
  title: string;
  /** queued | downloading | converting | done | error */
  status: string;
  /** 0-100 */
  progress: number;
  format: string;
  /** @nullable */
  outputPath?: string | null;
  /** @nullable */
  error?: string | null;
  createdAt: string;
  updatedAt: string;
}