# lib/api-zod/src/generated/types/libraryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AudibleBook } from "./audibleBook";

export interface LibraryResponse {
  books: AudibleBook[];
  total: number;
  page: number;
  pageSize: number;
}