# lib/api-zod/src/generated/types/libraryStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LibraryStats {
  total: number;
  downloaded: number;
  downloading: number;
  totalHours: number;
}